api_gateway/requirements.txt


Flask==2.2.3
Werkzeug==2.2.3
pymongo==4.3.3
Flask-PyMongo==2.3.0
PyJWT==2.6.0
python-dotenv==1.0.0
requests==2.28.2